#7126 create WidgetRenderer and move rendering widgets there #7126 some cleaning up
# Conflicts:
#	frontend/src/components/filters/Filters.js
