+++ frontend/src/index.jsx
@@ -5,7 +5,12 @@ import App from './containers/App';
 
 // if (process.env.NODE_ENV !== 'production') {
 //   const whyDidYouRender = require('@welldone-software/why-did-you-render');
-//   whyDidYouRender(React, { include: [/DocList/] });
+//   whyDidYouRender(React, { include: [/RawWidget/] });
+// }
+
+// if (process.env.NODE_ENV !== 'production') {
+//   const { whyDidYouUpdate } = require('why-did-you-update');
+//   whyDidYouUpdate(React, { include: [/RawWidget/] });
 // }
 
 /* eslint-disable */
