+++ frontend/src/reducers/windowHandler.js
@@ -521,7 +521,16 @@ export default function windowHandler(state = initialState, action) {
       } else {
         const currentVal = state[scope] ? state[scope][property] : {};
 
-        newValue = merge.recursive(true, currentVal, value);
+        // because some fields can have added/removed fields when there's
+        // a change and we don't want the leftovers in the store
+        if (property === 'data') {
+          newValue = {
+            ...currentVal,
+            ...value,
+          };
+        } else {
+          newValue = merge.recursive(true, currentVal, value);
+        }
       }
 
       return update(state, {
